コンポーネント分割見直し
