+++ src/components/atoms/SectionTitle.tsx
@@ -8,7 +8,7 @@ type Props = {
 export const SectionTitle: React.FC<Props> = ({ children }) => {
   return (
     <>
-      <Text size={26} component="h2" m={0}>
+      <Text size={26} component="h2" mt={40} mb={0}>
         {children}
       </Text>
       <Divider my="lg" pb={4} />
